Rename streamutils package to streamutil
