+++ streamutil/inmemstore/store.go
