+++ streamutil/inmemstore/store_test.go
